+++ src/ast/mod.rs
@@ -586,7 +586,9 @@ fn normalize_attribute_usage(a: &AttributeUsage) -> AttributeUsage {
         ordered: a.ordered,
         nonunique: a.nonunique,
         is_derived: a.is_derived,
+        usage_prefix: a.usage_prefix.clone(),
         is_constant: a.is_constant,
+        is_reference: a.is_reference,
         is_end: a.is_end,
         membership: a.membership.clone(),
     }
@@ -919,6 +921,9 @@ fn normalize_part_usage_body_element_node(
         PartUsageBodyElement::SuccessionUsage(n) => {
             PartUsageBodyElement::SuccessionUsage(dummy_node(n, n.value.clone()))
         }
+        PartUsageBodyElement::CalcUsage(n) => {
+            PartUsageBodyElement::CalcUsage(dummy_node(n, n.value.clone()))
+        }
         PartUsageBodyElement::Import(n) => {
             PartUsageBodyElement::Import(dummy_node(n, normalize_import(&n.value)))
         }
@@ -1164,6 +1169,7 @@ fn normalize_enum_def(e: &EnumDef) -> EnumDef {
 fn normalize_occurrence_def(o: &OccurrenceDef) -> OccurrenceDef {
     OccurrenceDef {
         is_abstract: o.is_abstract,
+        is_individual: o.is_individual,
         identification: o.identification.clone(),
         specializes: o.specializes.clone(),
         body: o.body.clone(),
@@ -1244,6 +1250,7 @@ fn normalize_end_decl(e: &EndDecl) -> EndDecl {
 
 fn normalize_ref_decl(r: &RefDecl) -> RefDecl {
     RefDecl {
+        direction: r.direction,
         name: r.name.clone(),
         type_name: r.type_name.clone(),
         typing: r.typing.clone(),
@@ -1259,6 +1266,7 @@ fn normalize_ref_decl(r: &RefDecl) -> RefDecl {
 
 fn normalize_action_def(a: &ActionDef) -> ActionDef {
     ActionDef {
+        is_individual: a.is_individual,
         identification: a.identification.clone(),
         specializes: a.specializes.clone(),
         body: normalize_action_def_body(&a.body),
@@ -1404,6 +1412,7 @@ fn normalize_action_usage(a: &ActionUsage) -> ActionUsage {
         is_abstract: a.is_abstract,
         is_variation: a.is_variation,
         is_reference: a.is_reference,
+        is_individual: a.is_individual,
         name: a.name.clone(),
         type_name: a.type_name.clone(),
         typing: a.typing.clone(),