+++ src/parser/payload.rs
@@ -169,6 +169,7 @@ fn control_node_payload_stmt<'a>(
                 is_abstract: false,
                 is_variation: false,
                 is_reference: false,
+                is_individual: false,
                 name: control_name.to_string(),
                 type_name: String::new(),
                 typing: None,